consumer/requirements.txt
confluent-kafka
requests
python-telegram-bot==13.15
python-dotenv
